Create Q2-PY
"Here is a record of me completing my homework alone:"
"the record time is November 6, 2024 at 19:48 pm, I completed "
"the second question and tested all the correct answers. "
"I uploaded this py file to my github to prove that this file was written by myself."
"If it is found that my code is highly overlapping with someone else's, " \
"there is no doubt that someone else has stolen my answers, and this PY file, " \
"along with GitHub history and time logs, will be used as evidence in " \
"the academic misconduct hearing. My name is Xinjie WU 学号20624223."